+++ lib/iiif_print.rb
@@ -2,6 +2,7 @@ require "iiif_print/engine"
 require "iiif_print/errors"
 require "iiif_print/jp2_image_metadata"
 require "iiif_print/image_tool"
+require "iiif_print/catalog_search_builder"
 require "iiif_print/text_extraction"
 require "iiif_print/data"
 require "iiif_print/configuration"
